refactor: Check the existence of schema.yml file

+++ contracts/wasm/scripts/cleanup.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 example_name=$1
 cd $example_name
+if [[ ! -f "schema.yaml" ]]; then
+    echo 'schema.yaml not exists'
+    exit
+fi
 schema -go -rs -ts -clean
 rm ./ts/"$example_name"impl/tsconfig.json
 rm ./rs/"$example_name"impl/Cargo.lock
